Added "rootOwner" notion to RMIRulesAgreementBuilder. Fixed unintended behavior rules described by BasicInvocationRule. Quick fix in MethodDescription to make accessible the java method. Made some optimisation for methods that does not describe any particular behavior for their parameters. Sync classes and the whole synchronized object system are now able to handle protected methods. MethodInvocation now contains the synchronizedObject's node instead of the synchronized object value. PersistenceConfigBuilder.setTConverter has been fixed (the objects being deserialized using this method where completely corrupted) ScalaUtils.setValue now able to set values to static fields

+++ Engine/src/main/scala/fr/linkit/engine/gnom/cache/sync/generation/SyncClassCompilationRequestFactory.scala
@@ -25,13 +25,15 @@ import fr.linkit.engine.internal.generation.compilation.factories.ClassCompilati
 import fr.linkit.engine.internal.mapping.ClassMappings
 
 import java.io.File
-import java.nio.file.Files
+import java.nio.file.{Files, NoSuchFileException}
 
 class SyncClassCompilationRequestFactory extends ClassCompilationRequestFactory[SyncObjectSuperclassDescription[_], SynchronizedObject[_]](DefaultClassBlueprint) {
 
     override def loadClass(req: CompilationRequest[Seq[Class[_ <: SynchronizedObject[_]]]], context: SyncObjectSuperclassDescription[_], className: String, loader: GeneratedClassLoader): Class[_] = {
-        val (byteCode, wrapperClass) = new ClassRectifier(context, className, loader, context.clazz).rectifiedClass
         val wrapperClassFile         = req.classDir.resolve(className.replace(".", File.separator) + ".class")
+        if (Files.notExists(wrapperClassFile))
+            throw new NoSuchFileException(s"class file for class $className at ${req.classDir} not found.")
+        val (byteCode, wrapperClass) = new ClassRectifier(context, className, loader, context.clazz).rectifiedClass
         Files.write(wrapperClassFile, byteCode)
         RuntimeClassOperations.prepareClass(wrapperClass)
         ClassMappings.putClass(wrapperClass)
